+++ src/routes/books/books.types.ts
@@ -1,0 +1,14 @@
+import { DefaultLoadingState } from "../../types";
+
+export interface Book {
+  id: string;
+  bookTitle: string;
+  bookAuthor: string;
+  bookDescription: string;
+  bookPrice: number;
+  bookImgUrl: string;
+}
+
+export type BooksState = DefaultLoadingState & {
+  books: Book[];
+}
