+++ src/routes/cart/cart.types.ts
@@ -1,0 +1,14 @@
+import { DefaultLoadingState } from "../../types";
+
+export type CartItem = {
+  id: string;
+  bookTitle: string;
+  bookAuthor: string;
+  bookDescription: string;
+  bookPrice: number;
+  bookImgUrl: string;
+}
+
+export type CartState = DefaultLoadingState & {
+  cart: CartItem[];
+};
