+++ src/routes/cart/cart.utils.ts
@@ -1,0 +1,3 @@
+import { CartItem } from "./cart.types";
+
+export const totalCartPrice = (cart: CartItem[]) => cart.reduce((total, item) => (total += item.bookPrice), 0);
