+++ src/types/types.ts
@@ -5,29 +5,4 @@ export type DefaultLoadingState = {
   error?: string;
 }
 
-export interface Book {
-  id: string;
-  bookTitle: string;
-  bookAuthor: string;
-  bookDescription: string;
-  bookPrice: number;
-  bookImgUrl: string;
-}
 
-export type BooksState = DefaultLoadingState & {
-  books: Book[];
-}
-
-
-export type CartItem = {
-  id: string;
-  bookTitle: string;
-  bookAuthor: string;
-  bookDescription: string;
-  bookPrice: number;
-  bookImgUrl: string;
-}
-
-export type CartState = DefaultLoadingState & {
-  cart: CartItem[];
-};
