+++ src/utils/utils.ts
@@ -1,4 +1,3 @@
-import { CartItem } from "../types";
 
 export const delay = (time: number): Promise<true> => {
   return new Promise((resolve) => {
@@ -7,5 +6,3 @@ export const delay = (time: number): Promise<true> => {
     }, time)
   })
 }
-
-export const totalCartPrice = (cart: CartItem[]) => cart.reduce((total, item) => (total += item.bookPrice), 0);
